Fixed bug#13625834 - main.innodb_mysql_lock test fails on debian5.0-x86_64                      sporadically.
The reason for this bug is that there was a race condition in checking for
'update' state in wait_condition loop and acquiring real transaction lock
during execution of statement 'INSERT'. Such race condition arised since
there was a time window between the step when server sets a transaction state
in 'update' value and when the real lock for transation is acquired. Such 
race condition could result in successful execution of the following statement
'INSERT INTO t1 VALUES (2)' that else should failed with error
ER_LOCK_DEADLOCK.

To eliminate this race condition we replaced the conditon of waiting for
lock by polling the table information_schema.innodb_trx until transaction will
be locked.