Fix Bug#13875080 STATS ESTIMATION MAY TRY TO DIVE INTO A NONEXISTING LEVEL ON A BTREE
Protect the whole scan of the index tree with an S latch.
Release and reacquire the latch now and then to avoid holding it
for too long time.

Introduce a new function btr_height_get() to fetch the height of a btree
and to release the lock acquired by btr_root_block_get() afterwards.

Change btr_root_block_get() to accept a new parameter telling it whether to
S or X lock the root page. This change is not really necessary for this fix
but is an easy enough optimization - no need to X lock the page in order to
peek at its level.

Change btr_cur_search_to_nth_level() to accept a new flag in latch_mode
(BTR_ALREADY_S_LATCHED) which tells it that the caller already has an S lock
on the index tree and that this function should not acquire/release it by
itself.

Cover the entire dict_stats_analyze_index() with one mtr that holds an S
lock on the index tree. Release and reacquire that lock (committing the mtr
and starting it again) before each n_prefix to avoid holding it for too long.
If we detect that the tree height has changed during the release/reacquire
then we quit the scan and leave the rest of the stats in whatever state they
were before dict_stats_analyze_index().

Approved by:	Marko (rb:992)