Bug#13701973 - 64274: MYSQL THREAD WAS SUSPENDED WHEN EXECUTE UPDATE QUERY
InnoDB FTS setup tried to follow foreign key references without keeping track of which ones it had
already seen. With circular and other complex setups, this could loop forever, or a very long time,
respectively, leading to the appearance of the query thread hanging.

This patch adds a depth failsafe.