Bug#12884336: Huge performance problem with subst_spvars and name_const formatting
In stored procedures, we generated loggable statements (with
variables replaced by NAME_CONST()) in a lot of cases where
it wasn't necessary (because nobody was using the resulting
string).

We now don't preemptively generate a loggable string if
a) general log is off, and
b) binary logging is off, or not in statement mode.

If slow query logging is active, we generate a loggable
statement only once we know the statement was actually slow
(unless of course we've already generated a loggable string
for any of the other logs, in which case we'll re-use that).