BUG#13877432: PERFORMANCE PROBLEM WHEN GTID IS ON
When writing GTIDs the server needs to rewind the position in the
CACHE and write the data at the beginning of the transaction cache. If
the IO_CACHE buffer is not big enough to accommodate all the changes
for the given transaction, it will swap data to disk and then a call
to reinit_io_cache is needed to have the buffer properly filled with
the correct data again. One cannot just use my_b_seek to rewind to the
beginning of the buffer/file.

Currently, flush_ and reinit_io_cache is called unconditionally every
time a GTID event is written, meaning that even in those cases that
all events from a transaction fit together in the IO_CACHE buffer a
flush to disk done.

This patch fixes this by only calling reinit_io_cache when it is
absolutely necessary (i.e., when the IO_CACHE actually swapped).

In addition, it also fixes the case where an unlock was missing when
write_empty_groups_to_cache was called but failed (there was a return
without unlocking the global_sid_lock.

Finally, the patch also make a small refactoring, reducing the amount
of code in gtid_before_write_cache.