WL#5223: Binary Log Group Commit
This is the main patch for implementing binary log group commit.

It performs a group flush to the binary log by maintaining a queue
of pending (session) threads that want to write to the binary log
and flush and commit them in batches.

This patch:

Introduces a simple lock-based thread queue with two functions:
- Enqueue a thread
- Fetch and empty the complete queue.

Introduces a class Thread_excursion to handle (session) thread excursions,
that is, attaching a POSIX thread temporarily to a session thread and
execute as if it was that session.

Extends the binary log caches with a "finalize" function that is
used to finalize the cache by writing the final COMMIT or ROLLBACK. After
a cache is finalized, no new events can be written to it. To reset the flag,
it is necessary to either flush or truncate the cache.

Introduce MYSQL_BIN_LOG::ordered_commit which is called from the function
MYSQL_BIN_LOG::commit and MYSQL_BIN_LOG::rollback to execute a ordered group
flush and commit.