WL#6168: FLUSH TABLES ... FOR EXPORT -- parser
This WL introduces the new "FLUSH TABLES ... FOR EXPORT" syntax for the
larger WL5522.

The "FOR EXPORT" option is mutually exclusive with the old
"WITH READ LOCK" lock.

A temporary implementation for the "FLUSH TABLES .... FOR EXPORT" command
just copies the behaviour of "FLUSH TABLES ... WITH READ LOCK".

Further logic of the "FOR EXPORT" lock will be implemented by the WL6169.