Fix for bug#13901890 SQL NODE CRASHES DURING INSERT SELECT
There was a bug in the lifetime handling of NdbQuery objects
where the 'root' of the pushed join had 'type == const'

As part of integrating the pushed join feature into the 
ha_ndbcluster handler implementation, the default implementation
of handler::index_read_idx_map() was redefined. This method is
used as the access function to read the (single) row when
a table is const'ified by the optimizer.

Basically handler::index_read_idx_map() does (ignoring error handling):

  index_init(index, 0);  // Open table for access
  index_read_map();      // Virtual method reading single row
  index_end();           // Return table to 'closed' state + release

As the above 'index_end' would also destruct the resultset from
the pushed join, including the child result, we could not do that 
for a pushed join.

Therefore, ha_ndblcluster::index_read_idx_map() was implemented *without*
the final index_end() call - Our investigation at that time indicated that
this could be omitted as the open tables was cleaned up anyway, either by the
next operation on the handler instance, or when the NdbTransaction was terminated.

However, this bugs uncovers that there are codepaths where we both:

- Terminate NdbTransaction, which destruct the NdbQuery object.
- Then call ha_ndbcluster::reset() to clean up the handler, which will
  result in NdbQuery::close() to be called -> Crash!

Working with this problem I realized that:

- In order to increase pushability we have already introduced the 
  handler::test_push_flag(HA_PUSH_BLOCK_CONST_TABLE) inside the optimizer
  which will block the const'ifying during optimize.
- It will then only be confusing to still explain these query plan
  having a type == 'const' as they are now really executed as an eq_ref
- Override ::index_read_idx_map() to suite pushed join execution is
  obsolete if we instead handle these queries as an eq_ref.

These changes does *not* affect the pushability of the queries, nor
changes number of handler call required to retrieve the data.

However, several EXPLAINS will change where 'type == const' will
now show 'type==eq_ref' instead which IMHO is more correct.

Also fixes incorrect destruction of NdbQueryImpl objects if 
construction of these fails in NdbQueryImpl::buildQuery().
In order to ensure propper cleanup of these we should use 
NdbQueryImpl::release() instead of deleting them directly.