Fix Bug#13946118 STATS RELATED, ASSERT (RW_LATCH == 1) || (RW_LATCH == 2) || (RW_LATCH == 3)
This bug was introduced in vasil.dimov@oracle.com-20120322170754-3mdqjdq93nuzqb04

In btr_pcur_open_low() we shold clear the newly added flag
BTR_ALREADY_S_LATCHED from latch_mode because later in buf_page_get_gen()
we assert that latch_mode is one of RW_S_LATCH, RW_X_LATCH or RW_NO_LATCH.

This was discussed with Marko and is fixed in WL#6175 (which is not pushed
yet). I am not sure why it was fixed there instead of mysql-trunk, since
the problem exists in mysql-trunk.