WL#5223: Binary Log Group Commit
This patch split the commit into several phases
to allow several threads to work in parallel on
committing transactions. Each phase is run by a
leader, which will execute a commit on behalf of
zero or more followers. If a leader catches up on
the preceeding phase, it will merge with that set
and allow the leader for that phase to run the
commits for itself and all its followers. The
intention is to allow the commit procedure to be
self-adapting, so that slow phases will accumulate
larger batches of commits to execute.

The patch also extends the build files to gather
information about the CPU (specifically, the cache
line size).