Bug #13966514: CRASH IN GET_SCHEMA_TABLES_RESULT WITH MIN/MAX, LEFT/RIGHT JOIN ON I_S TABLE
When there's a MIN/MAX that's reduced to a constant due to the 
presence of an unique index over the aggregated column the
tables from the query are optimized away.
If one of these removed tables happens to be a I_S table 
the materialization check will still run due to the fact that the 
the select_lex flag OPTION_SCHEMA_TABLE is not cleared up.
Fixed by adding a check in get_schema_tables_result() for the 
presence of join tables (join_tab) before an attempt is made to 
iterate over them (which leads to a null pointer dereference crash).
Test suite added.