WL#5545 test coverage improvements.
Test if incompletely created indexes will be dropped when there is
a table name#2 clash when creating a clustered index.

ha_innobase::commit_inplace_alter_table(): Fix broken error handling.
Do not decrement MONITOR_PENDING_ALTER_TABLE if an error is being
returned for commit=true. In this case, the method will be re-invoked
with commit=false. This bug was introduced in WL#5526.