Fix for Bug#13990924 '64865: LARGE WHERE IN WITH SPJ LEADS TO CLUSTER SHUTDOWN'
As part of developing the SPJ feature, the macro DEBUG_CRASH() was used
several places where an error occured. I assume the intention
has been to only (permanently) use this macro in places where we
did not handle the error gracefully.

However, use of this macro has been exaggerated such that it is
even used when we catch API errors in the SPJ block which we
*do* handle by returning an error code.

Furthermore this macro was defined as ndbrequire(false), such that
we always crash when we hit it - even in GA code.

For this particular bug, the SPJ block detected the condition:

     err = DbspjErr::QueryNodeParametersTooBig;
     if (unlikely(param_len >= NDB_ARRAY_SIZE(m_buffer1)))
     {
       DEBUG_CRASH();
       goto error;
     }

Furthermore, there was a check in the SPJ API which was intended to
prevent sending 'too big' parameters to the SPJ node. However, this check
was 'size <= 64k', while the actual size limit in SPJ block was 32K!

This fix:

 - Redefined DEBUG_CRASH to an ndbassert(false) which will not fail
   in production code.

 - Replaced lots of these DEBUG_CRASH() with a jam() in places where
   we do handle the error.

 - Increase buffer sizes in SPJ block from 32K -> 64K.