BUG#13733221 - 64345: MYSQLD CRASHES ON SHUTDOWN WHEN SLOW LOG IS A PIPE
We could not properly log to a named pipe as logging did not handle failing
ftell() well. We now just ignore it if log-file is a pipe.

A related error during shutdown tried to print a per-thread message without
having a thd; we now use the default messages as we should.