BUG#13983918 - FACTOR OUT MY_GET_ERR_MSG() FROM MY_ERROR() FOR GENERAL USE
In order to use the format strings in sql/share/errmsg-utf8.txt we should
factor out the format string lookup code into a separate function. That way
we can use the function to write to the error log and also to push
info/warning message to the client using the format string from
sql/share/errmsg-utf8.txt. This is required for WL#5522 to work.