Fix for Bug#13810048 INCORRECT JOIN ORDER USED DUE TO INCORRECT                      COST CALCULATIONS
The problem is that we fail to select the best join order for a two
table join. The cause for the wrong join order is that
best_access_path() selects the wrong index to use for the join. The
selected index has much higher record and cost estimate than the best
index. Because of the high index cost, the estimated cost of the optimal
join order became very high, and another, less optimal, join order was
selected.

Assume best_access_path() can select among the following indexes:

  (a, d)
  (b)
  (b, c)

and that the select condition is:

  a=1 AND b = 1 AND c <> 3

The current code in best_access_path() will compute record and cost
estimates for each of these. What causes the bug to happen is an
heuristic introduced as a fix for Bug#26106. This heuristic was
introduced to ensure that when considering the index (b, c) this
should be selected based on that this index has a quick select object
and this quick select object is able to use more parts of the index
than the other alternative indexes. The code that does this is:

  /* 
    Fix for the case where the index statistics is too
    optimistic: If 
    (1) We're considering ref(const) and there is quick select
        on the same index, 
    (2) and that quick select uses more keyparts (i.e. it will
        scan equal/smaller interval then this ref(const))
    (3) and E(#rows) for quick select is higher then our
        estimate,
        Then 
          We'll use E(#rows) from quick select.

    Q: Why do we choose to use 'ref'? Won't quick select be
    cheaper in some cases ?
    TODO: figure this out and adjust the plan choice if needed.
  */
  if (!found_ref && table->quick_keys.is_set(key) &&    // (1)
      table->quick_key_parts[key] > max_key_part &&     // (2)
      records < (double)table->quick_rows[key])         // (3)
  {
    records= quick_records= (double)table->quick_rows[key];
    quick_matches_more_parts= true;                  
  }

Setting quick_matches_more_parts to true will ensure that this index is
select even if it has a much higher cost than the currently best
index. The (a, d) and (b) index can only use the first part of the index
while (b, c) can use two parts of the index. So the code will select
to use (b, c) without considering the cost of using (a, d) or
(b). This is wrong, since the (a, d) index might select a lot fewer
records even using only the first part of the index.

The goal of the fix for Bug#26106 was to solve the cases where we
"wrongly" selected to use (b) instead of (b, c). Using (b, c) should
always be equal or less costly than using (b) since it will produce
fewer (or the same) number of records (if we ignore that it is more
costly to use an index with more columns.).

The fix for this problem is to revert the patch for Bug#26106. This
will re-introduce the "bug" that we can select to use an index with
with few fields (e.g. (b)) when there is another index, (e.g. (b, c))
that can use more parts of the index and thus will produce the same or
fewer records. The argument for allowing this is that this situation
should only occur when the record and cost estimates for these two
indexes are very close and thus the cost of using either of these are
almost similar. Basically, by reverting the patch we return to let the
choice of using (b) or (b, c) be cost based instead of based on an
extra heuristic.

The assumption is that this can now only happen when the real cost of using
(b) or (b, c) is similar. If this assumption is wrong then we need to
first understand why we choose the wrong index or why the cost numbers
are wrong before making an alternative fix. This might either be to
fix the cost numbers or implement a better version of the heuristic to
always select the index that can use most parts of the index when
another index with a common prefix is the best index based on cost
estimates.