Extend the fix for Bug#13838962 STATISTICS CALCULATION SKIPS DELETE-MARKED RECORDS AND CAUSES ASSERTION FAILURE
Counting all delete marked records brings too much nondeterminism because
purge may or may not have wiped them away when table and index statistics
are being gathered.

Instead skip delete marked records only on the leaf pages - that is, behave
as if they do not exist. Delete marks on records on non-leaf pages are
nonsensical and should not be looked up at all - that is, behave as if
the non-leaf records do not _contain_ a delete-marked bit.