WL#6255 preparation: Suspend the purge if it would free BLOBs.
We cannot free BLOBs during online table rebuild, because we do not write
the BLOBs to the online log. Online log apply assumes that the BLOBs are
accessible.

row_purge_upd_exist_or_extern_func(), row_purge_record_func(): Return true
if the record was purged, false if it would free BLOBs of a table that is
being rebuilt online.

row_purge(): If a record is not purgeable, retry in 1 second, until the
server is shut down is initiated. TODO: move the record aside, and continue
with the rest of the purge.