Bug #13890837: DEFAULT() CALL MAKES MYSQL SERVER CRASH
PROBLEM:
We create a item of type DEFAULT_VALUE_ITEM to represent 
DEFAULT(b) in the query. While creating a temporary 
table, we do not create any field for this item as it 
is a const_item. 
In the call to remove_duplicates(this is where we dump), 
we check if the field is a const_item. If not, we access 
the contents of the field item in the temporary table.

In this case,  the call to const_item() inside 
remove_duplicates() returns false, as we would have 
changed the type of the item in the call to 
get_tmp_table_item() inside change_refs_to_tmp_fields(). 
As a result, we try to check the contents of the field 
and dump.

As class Item_default_value does not have its own 
get_tmp_table_item function, it calls 
Item_field::get_tmp_table_item(). This results in the 
change of the type of the object created and therefore
the dump.

FIX:
Introduce a call to get_tmp_table_item() for class 
Item_default_value, so that the type of the object 
remains same.