Bug#13903821 WL#5526 performance regression in row_ins()
Add a trx_id field to row_prebuilt_t and track when the last insert query
graph was created. If the last index in the list has a higher trx id then
recreate the insert query graph.

This also fixes a memory leak that was introduced as part of online index
worklog. The memory leak caused the regression too.

rb://1097 Approved by Marko