Bug#14001091 ASSERT IN DIAGNOSTICS_AREA::SET_EOF_STATUS              DURING ANALYZE TABLE
This assertion could happen if internal commits during a table
administration statement (e.g. ANALYZE TABLE) failed (e.g. due
to lock timeout). The result of the commit was not checked so
the failure was not detected. When the statement at the end
of execution tried to report 'OK' the assertion was triggered.
The assertion checks that we're not trying to report 'OK'
after having reported an error earlier. This bug had no
consequences on release builds, there the error would simply
be kept.

This patch fixes the problem by checking if the commit has in
fact completed successfully. If so, the statement won't try
to report 'OK' to the client.