Bug#13974815 MULTITABLE UPDATE IGNORE WITH FAILING FK CONSTRAINT:              ASSERT 0 IN PROTOCOL.CC:518
This bug could be triggered by multi update with ignore
if there was some kind of handler error (e.g FK constraints).
The consquence on release builds was that it was possible for
'OK' to be sent to the client even if the statement had in
fact failed.

The problem was that ignore caused lex->current_select->no_error
to be set, which caused the handler error to be converted to
a warning. So, while the update failed, no error was reported -
triggering the assert. This should not have been possible since
the handler error was determined to be fatal (see
handler::is_fatal_error()), however the flag indicating that the
error was fatal was not propagated...

This patch fixes the problem by making sure that handler error
flags (such as ME_FATALERROR) are properly propagated by
handler::print_error().

Note that this patch adds an extra parameter to two handler API
functions. However, both functions have already been changed 
in earlier 5.6 patches (WL#5534).