BUG#12980522 - MULTI-UPDATE REFERENCING VIEW IS BLOCKED BY                --READ-ONLY WITHOUT GOOD REASON
BACKGROUND:
Multi-update which references view in its main JOIN but
doesn't update it, can be blocked by --read-only/@@read_only=1
mode even though it updates only temporary tables. This should
not be happening (and is not happening if the view is replaced
by normal table).

ANALYSIS:
In mysql_multi_update_prepare() function, updating variable
is correctly set to 0 for all non-temporary tables that are
not getting updated. 
But for view this is not happening. This value is not correctly
set for views. 

FIX:
This bug is fixed by correctly setting the updating variable to
0 if the table is a view as it is correctly set for the tables
comprising the view.