Bug #14169459 INNODB; DROP TABLE DOES NOT DELETE THE IBD FILE FOR A TEMPORARY TABLE.
This patch has a number of changes. The main one to fix this problem
is an extra test in row_drop_table_for_mysql() for '!temp' before
calling fil_space_for_table_exists_in_mem() to check for 'unknown or
discarded' tablespaces.

The patch also adds the ibd extension to the temporary tablespace path
before creating a fil_space_t.  This is how 5.5 does it.  

Another change is in fil_delete_file(), a new function associated with
the worklog 5522, where fil_make_cfg_name() and fil_make_ibt_name() are
called with a tablename instead of a path as they should.
This caused the function to not work at all for cfg and ibt files.
Because of this, innodb-index-online.test was leaving a 't1.ibt' file
behind in the data directory.  

Also, in fil_delete_tablespace(), where it is called with rename=true,
the logic was slightly wrong so that if an ibt file already existed, the
code would fail to rename the ibd tablespace to the ibt extention.

The function os_file_delete_if_exists() is improved so that if a directory
does not exist on Windows, it behaves just a s if the file does not exist.

Some related functions were changed to use bool instead of ibool.
And row_drop_table_for_mysql() was cleaned up in its use of multiple
tablename variables.