Attempt to fix several intermittent and platform specific PB2 failures from the wl 5980 push.
innodb-multiple-tablespaces;  This test causes a code path where
InnoDB will not start up because there are two versions of a
tablespace available during recovery.  This path should exit with
exit(1) instead of ut_a() since MTR will stop the test if it finds
a call stack on some platforms.

This test had a very rare timing related error where a --remove-file 
command in MTR did not work. This patch replaces --copy_file followed
by --remove_file with just a single --move-file.

innodb-wl5980-discard;  One failure on one platform  out of over a
dozen runs involves a pre-existing entry in the InnoDB tablespace
cache for 'test/t1' when a new 'test/t1' is being added.  In case
this is  a conflict with another test run in parallel, the tablename
is changed to 'test/t5980'.