Fix for Bug#14053325 DUPLICATED RESULTS ROWS WHEN SEMIJOIN USE LOOSESCAN STRATEGY
'Semijoin LooseScan strategy' is based on skiping key duplicates after the
'first match' for each key is found. Thus reading of the tables has to return
the result in sorted order as defined by the index being used.
      
Such 'sorted' read is specified with the 'sorted' argument to handler::ha_index_init().
As we failed to enforce 'sorted=true' for SJ_OPT_LOOSE_SCAN, it is left to each SE-handler
to return the rows sorted or unsorted. 
      
Partitioned tables and ndbcluster tables will not be sorted without 'sorted' being
set to 'true'. The skipreading in LooseScan will therefore fail to correctly eliminate
duplicates, and we get to many rows in the result set.