WL#6255 rb:1105 Problem 26 fix.
trx_undo_prev_version_build(): Replace the assertion on table_id
mismatch with a check. On table_id mismatch, assert that the table has
been rebuilt by a transaction that started after the record version
was written.

This should be repeateable by a single connection thread, with enough
purge lag:

create table;
lots_of_updates;
alter table;
begin;
do_something;
rollback; -- crashes

Before WL#6255, we rebuilt the table using a DB_ROLL_PTR that looks
like an insert. Those result in a DB_SUCCESS return from
trx_undo_prev_version_build().

WL#6255 is copying the DB_TRX_ID,DB_ROLL_PTR from the old table as is,
so that it can detect which changes were already applied from the
online rebuild log.

The problem occurs when DB_ROLL_PTR is an update_undo record, and when
the purge view has not yet advanced to the DB_TRX_ID that is stamped
on the record. The undo log record is still valid, but it is pointing
to the old table. Because the rebuild did not copy any history, and
because the rolling-back transaction started after the rebuild, it is
safe to pretend that the history was already purged. The purge will
eventually happen. It will read the undo log records, figure out that
the table_id has been dropped, and remove the undo log records without
touching any tables.

Approved by Jimmy Yang on IM.