This is a fix for bug#14208924 . It is based on an initial patch written by Frazer Clement.
If TC decides to abort a transaction in the 'prepared' state, it may leak 
ApiConnectRecord objects that was allocated via Dbtc::seizeApiConnectCopy(). 
These are normally released by Dbtc::sendApiCommit() followed by
Dbtc::releaseApiConCopy(). If the transaction aborts, these methods are not 
called.

This patch fixes the above issue by checking if the ApiConnectRecord object
has an apiCopyRecord when the transaction aborts. If so, the copy object is put
back in the free list.

The patch also adds a regression test case (using an error insert).