Bug#14061851 SERVER DDL: ASSERTION FAILED: ! IS_SET(),              FILE SQL_ERROR.CC, LINE 436
The assertion could be triggered by CREATE/ALTER/DROP SERVER.
The reason was that error handling was broken. One problem
was that callers generally assumed that callees didn't report
their own errors when they in fact did. Another problem was
that the error returned from SERVER functions was assumed to
have just one argument when they in fact could have several.
A third problem was that when open_ltable() failed, it was
assumed that my_errno contained an appropriate error.
A forth problem was that the possibility of KILL QUERY/
KILL CONNECTION during SERVER statements was not taken into
account.

This patch fixes the problem by making callees responsible
for reporting errors (like is generally done in the server).