bug#14092635 MTS slave risks consistency when event updates FK-related tables
This patch is for review.

MTS Scheduler can't account implicitly involved through FK-dependencies databases.
That can lead to (temporary) loss of data consistency on the slave.

Fixed with marking replication events on the master so MTS schedules them 
sequentially to avoid races and thereby inconsistency.

Solution is bound to be performance restrictive until WL_6049 has provided
interfaces to the table FK-graph.