Bug#14075016: Extra rows returned on WHERE ... IN subquery with nested joins
In this bug, we have a MaterializeScan semi-join operation and an OR
condition that is not pushed down to a table in the join plan.
This means that some non-matching rows are not filtered out, and the
query returns 20 instead of the expected 18 rows.
I tried to eliminate the derived table from the test case, but then
another plan was always selected.
Here is the join plan (materialization of derived table not shown)
  table3
  table1
  subquery2_t1 (MaterializeScan first)
  subquery2_t2
  subquery2_t3 (MaterializeScan last)
  table2

Here is the filter condition, with multiple equalities:
  ((table1.col_varchar_nokey <> 'm' or
   =('p', table2.col_varchar_key, subquery2_t1.col_varchar_nokey)) and
  =(table2.col_varchar_key, subquery2_t1.col_varchar_nokey) and
  =('1', table2.col_int_key) and
  =(subquery2_t3.col_varchar_key, subquery2_t2.col_varchar_key,
    subquery2_t3.col_varchar_nokey))

We focus on the multiple equality inside the OR branch, ie
  =('p', table2.col_varchar_key, subquery2_t1.col_varchar_nokey).

We also notice that there is an upper-level multiple equality that
partially covers this:
  =(table2.col_varchar_key, subquery2_t1.col_varchar_nokey).

Thus, when eliminate_item_equal() converts the multiple equality into
equality predicates, it generates subquery2_t1.col_varchar_nokey='p'
but not table2.col_varchar_key='p'.

After elimination, we get this generated condition:
  subquery2_t3.col_varchar_key = subquery2_t2.col_varchar_key and
  subquery2_t3.col_varchar_nokey = subquery2_t2.col_varchar_key and
  table2.col_int_key = '1' and
  table2.col_varchar_key = subquery2_t1.col_varchar_nokey and
  (table1.col_varchar_nokey <> 'm' or subquery2_t1.col_varchar_nokey = 'p')

Now, there is a problem with the OR condition. For the join_tab for table1,
subquery2_t1 is not available, and for subquery2_t1, table1 is not
available (it is outside of the materialized subquery). Thus, the first
place where this condition may be evaluated is table2. However, the way
we add used_table information in make_join_select() makes us assume that
this condition has already been pushed to a table, so it is not considered.

The simplest solution seems to be to make sure that the predicate
table2.col_varchar_key='p' is also added to the condition. After this,
we get this condition attached to table2:

  table2.col_int_key = '1' and
  (table1.col_varchar_nokey <> 'm' or
   (subquery2_t1.col_varchar_nokey = 'p' and table2.col_varchar_key = 'p'))

How do we do it: If an item from an outer table is eliminated because
it is equivalenced with an item from a materialized semi-join nest and
it is covered by an upper-level multiple equality, cancel the elimination.

internal/mysql-test/suite/i_main/t/subquery.test
  Test case for bug#14075016

internal/mysql-test/suite/i_main/r/subquery.result
  Test case results for bug#14075016.

mysql-test/r/subquery_sj_all.result
mysql-test/r/subquery_sj_all_bka.result
mysql-test/r/subquery_sj_all_bka_nixbnl.result
mysql-test/r/subquery_sj_all_bkaunique.result
mysql-test/r/subquery_sj_mat.result
mysql-test/r/subquery_sj_mat_bka.result
mysql-test/r/subquery_sj_mat_bka_nixbnl.result
mysql-test/r/subquery_sj_mat_bkaunique.result
  There are some explain changes because some predicates have been
  added to existing tests.

sql/sql_optimizer.cc
  In eliminate_item_equal(), added another criterion for keeping
  a predicate that was rejected because it was covered by an
  upper-level multiple equality.