Bug#13359247 ASSERT "LENGTH <= SHARE->NORMALIZED_PATH.LENGTH"              DURING OPEN_TABLE_DEF()
This assert could be triggered if opening of the FRM file for
a temporary table failed, e.g. because the FRM did not exist.

If opening of a FRM file fails, we try to unencode the file
name for 5.0 compatibility. However, the process of constructing
this file name fails to take temporary tables into account.
(i.e. temporary table FRMs are stored in a separate directory
and FRM file names are not constructed from the table name).
This caused the assertion to be triggered for temporary tables.

This patch fixes the problem by skipping the unencoding process
for temporary tables since such tables are not persistent anyway.